Change sp_depth to u32.

+++ src/backend.rs
@@ -112,7 +112,7 @@ pub struct Context {
     regs: Registers,
     /// Offset from starting value of SP counted in words. Each push and pop 
     /// on the value stack increments or decrements this value by 1 respectively.
-    sp_depth: usize,
+    sp_depth: u32,
 }
 
 impl Context {
